Restructure rest service folders
