+++ .eslintignore
@@ -1,2 +1,2 @@
-lib/cover-service-api/model/
-lib/cover-service-api/cover-service.ts
+lib/rest/cover-service-api/generated/
+lib/rest/cover-service-api/generated/cover-service.ts
