+++ components/shared/workCard/WorkCard.tsx
@@ -2,13 +2,13 @@ import Image from "next/image"
 import Link from "next/link"
 import React from "react"
 
-import { useGetCoverCollection } from "@/lib/cover-service-api/cover-service"
-import { GetCoverCollectionSizesItem } from "@/lib/cover-service-api/model"
 import { WorkTeaserFragment } from "@/lib/graphql/generated/fbi/graphql"
 import { getRandomContentColorClass } from "@/lib/helpers/colors"
 import { cn } from "@/lib/helpers/helper.cn"
 import { getIsbnsFromWork } from "@/lib/helpers/ids"
-import { useGetV1ProductsIdentifier } from "@/lib/publizon-api/publizon"
+import { useGetCoverCollection } from "@/lib/rest/cover-service-api/generated/cover-service"
+import { GetCoverCollectionSizesItem } from "@/lib/rest/cover-service-api/generated/model"
+import { useGetV1ProductsIdentifier } from "@/lib/rest/publizon-api/generated/publizon"
 
 import { Badge } from "../badge/Badge"
 import Icon from "../icon/Icon"
