+++ components/shared/workCard/helper.ts
@@ -1,11 +1,11 @@
 import { flatten } from "lodash"
 
-import { Cover } from "@/lib/cover-service-api/model"
 import { SearchWithPaginationQuery, WorkTeaserFragment } from "@/lib/graphql/generated/fbi/graphql"
 import { filterFalsyValuesFromArray } from "@/lib/helpers/arrays"
 
-import { CoverImageUrls } from "./../../../lib/cover-service-api/model/coverImageUrls"
 import { GeneralMaterialTypeCode } from "./../../../lib/graphql/generated/fbi/graphql"
+import { Cover } from "./../../../lib/rest/cover-service-api/generated/model"
+import { CoverImageUrls } from "./../../../lib/rest/cover-service-api/generated/model/coverImageUrls"
 
 export const displayCreators = (
   creators: SearchWithPaginationQuery["search"]["works"][0]["creators"],
