+++ package.json
@@ -11,8 +11,8 @@
     "test:unit": "vitest",
     "codegen:graphql": "graphql-codegen --config codegen.ts",
     "codegen:all-rest-services": "orval",
-    "codegen:covers": "rm -rf src/core/cover-service-api/model/*.* && orval --project coverService",
-    "codegen:publizon": "rm -rf lib/publizon-api/model/*.* && orval --project publizonAdapter"
+    "codegen:covers": "rm -rf src/lib/rest/cover-service-api/generated/model/*.* && orval --project coverService",
+    "codegen:publizon": "rm -rf lib/rest/publizon-api/generated/model/*.* && orval --project publizonAdapter"
   },
   "dependencies": {
     "@next/env": "^14.2.15",
