+++ tsconfig.json
@@ -30,7 +30,8 @@
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": [
     "node_modules",
-    "lib/cover-service-api/model",
-    "lib/cover-service-api/cover-service.ts"
+    "lib/rest/cover-service-api/model",
+    "lib/rest/cover-service-api/cover-service.ts",
+    "lib/graphql/generated/fbi/*.tsx"
   ]
 }
